+++ packages/components/input/src/textarea.tsx
@@ -1,11 +1,12 @@
-import {dataAttr} from "@heroui/shared-utils";
+import type {UseInputProps} from "./use-input";
+
+import {dataAttr, mergeProps} from "@heroui/shared-utils";
 import {forwardRef} from "@heroui/system";
-import {mergeProps} from "@react-aria/utils";
 import {useMemo, useState} from "react";
 import TextareaAutosize from "react-textarea-autosize";
 import {CloseFilledIcon} from "@heroui/shared-icons";
 
-import {UseInputProps, useInput} from "./use-input";
+import {useInput} from "./use-input";
 
 type NativeTextareaProps = React.TextareaHTMLAttributes<HTMLTextAreaElement>;
 type TextareaAutoSizeStyle = Omit<