+++ packages/components/input/stories/input.stories.tsx
@@ -1,9 +1,10 @@
 /* eslint-disable jsx-a11y/interactive-supports-focus */
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import type {ValidationResult} from "@react-types/shared";
+import type {Meta} from "@storybook/react";
+import type {InputProps} from "../src";
 
 import React from "react";
-import {Meta} from "@storybook/react";
 import {input} from "@heroui/theme";
 import {
   MailFilledIcon,
@@ -16,7 +17,7 @@ import {button} from "@heroui/theme";
 import {useForm} from "react-hook-form";
 import {Form} from "@heroui/form";
 
-import {Input, InputProps, useInput} from "../src";
+import {Input, useInput} from "../src";
 
 export default {
   title: "Components/Input",
@@ -50,7 +51,7 @@ export default {
       control: {
         type: "select",
       },
-      options: ["inside", "outside", "outside-left"],
+      options: ["inside", "outside", "outside-left", "outside-top"],
     },
     isDisabled: {
       control: {
@@ -175,15 +176,16 @@ const LabelPlacementTemplate = (args) => (
   <div className="w-full flex flex-col items-center gap-12">
     <div className="flex flex-col gap-3">
       <h3>Without placeholder</h3>
-      <div className="w-full max-w-xl flex flex-row items-end gap-4">
+      <div className="w-full flex flex-row items-end gap-4">
         <Input {...args} description="inside" />
         <Input {...args} description="outside" labelPlacement="outside" />
         <Input {...args} description="outside-left" labelPlacement="outside-left" />
+        <Input {...args} description="outside-top" labelPlacement="outside-top" />
       </div>
     </div>
     <div className="flex flex-col gap-3">
       <h3>With placeholder</h3>
-      <div className="w-full max-w-xl flex flex-row items-end gap-4">
+      <div className="w-full flex flex-row items-end gap-4">
         <Input {...args} description="inside" placeholder="Enter your email" />
         <Input
           {...args}
@@ -195,6 +197,12 @@ const LabelPlacementTemplate = (args) => (
           {...args}
           description="outside-left"
           labelPlacement="outside-left"
+          placeholder="Enter your email"
+        />
+        <Input
+          {...args}
+          description="outside-top"
+          labelPlacement="outside-top"
           placeholder="Enter your email"
         />
       </div>