+++ packages/components/pagination/src/use-pagination-item.ts
@@ -1,11 +1,11 @@
 import type {Ref} from "react";
 import type {HTMLHeroUIProps, PropGetter} from "@heroui/system";
 import type {LinkDOMProps, PressEvent} from "@react-types/shared";
+import type {PaginationItemValue} from "@heroui/use-pagination";
 
 import {useMemo} from "react";
-import {PaginationItemValue} from "@heroui/use-pagination";
-import {clsx, dataAttr} from "@heroui/shared-utils";
-import {chain, mergeProps, shouldClientNavigate, useRouter} from "@react-aria/utils";
+import {shouldClientNavigate, useRouter} from "@heroui/system";
+import {clsx, dataAttr, chain, mergeProps} from "@heroui/shared-utils";
 import {filterDOMProps, useDOMRef} from "@heroui/react-utils";
 import {useHover, usePress} from "@react-aria/interactions";
 import {useFocusRing} from "@react-aria/focus";