+++ packages/components/tooltip/src/tooltip.tsx
@@ -1,13 +1,14 @@
+import type {UseTooltipProps} from "./use-tooltip";
+
 import {forwardRef} from "@heroui/system";
 import {OverlayContainer} from "@react-aria/overlays";
 import {AnimatePresence, m, LazyMotion} from "framer-motion";
 import {TRANSITION_VARIANTS} from "@heroui/framer-utils";
-import {warn} from "@heroui/shared-utils";
+import {warn, mergeProps} from "@heroui/shared-utils";
 import {Children, cloneElement, isValidElement} from "react";
 import {getTransformOrigins} from "@heroui/aria-utils";
-import {mergeProps} from "@react-aria/utils";
 
-import {UseTooltipProps, useTooltip} from "./use-tooltip";
+import {useTooltip} from "./use-tooltip";
 
 export interface TooltipProps extends Omit<UseTooltipProps, "disableTriggerFocus" | "backdrop"> {}
 
@@ -54,7 +55,7 @@ const Tooltip = forwardRef<"div", TooltipProps>((props, ref) => {
 
       trigger = cloneElement(child, getTriggerProps(child.props, childRef));
     }
-  } catch (error) {
+  } catch {
     trigger = <span />;
     warn("Tooltip must have only one child node. Please, check your code.");
   }