+++ packages/core/theme/src/components/input.ts
@@ -61,7 +61,6 @@ const input = tv({
       "outline-hidden",
       "select-none",
       "opacity-0",
-      "hover:!opacity-100",
       "cursor-pointer",
       "active:!opacity-70",
       "rounded-full",
@@ -143,13 +142,13 @@ const input = tv({
       md: {
         inputWrapper: "h-10 min-h-10 rounded-medium",
         input: "text-small",
-        clearButton: "text-large",
+        clearButton: "text-large hover:!opacity-100",
       },
       lg: {
         label: "text-medium",
         inputWrapper: "h-12 min-h-12 rounded-large",
         input: "text-medium",
-        clearButton: "text-large",
+        clearButton: "text-large hover:!opacity-100",
       },
     },
     radius: {
@@ -178,6 +177,10 @@ const input = tv({
         inputWrapper: "flex-1",
         mainWrapper: "flex flex-col",
         label: "relative text-foreground pe-2 ps-2 pointer-events-auto",
+      },
+      "outside-top": {
+        mainWrapper: "flex flex-col",
+        label: "relative text-foreground pb-2 pointer-events-auto",
       },
       inside: {
         label: "cursor-text",